unix/windows: pulseaudio optional
