+++ CMakeLists.txt
@@ -2,6 +2,7 @@ cmake_minimum_required(VERSION 3.14)
 
 project(ASCIIpatrol LANGUAGES CXX)
 
+option(sound "play sound (Unix, Windows)" true)
 option(web "build WASM using emscripten")
 
 # --- auto-ignore build directory
@@ -22,8 +23,11 @@ elseif(UNIX)
   find_package(Threads REQUIRED)
   find_library(XI_LIBRARY NAMES Xi REQUIRED)
   find_path(XI_INCLUDE_DIR NAMES X11/extensions/XInput2.h REQUIRED)
-  find_library(PULSEAUDIO_LIBRARY NAMES pulse REQUIRED)
-  find_path(PULSEAUDIO_INCLUDE_DIR NAMES pulse/pulseaudio.h REQUIRED)
+
+  if(sound)
+    find_library(PULSEAUDIO_LIBRARY NAMES pulse REQUIRED)
+    find_path(PULSEAUDIO_INCLUDE_DIR NAMES pulse/pulseaudio.h REQUIRED)
+  endif()
 endif()
 
 # --- main program
@@ -50,15 +54,19 @@ elseif(CMAKE_CXX_COMPILER_ID STREQUAL OpenWatcom)
   target_sources(asciipat PRIVATE spec_dos.cpp)
   target_compile_options(asciipat PRIVATE -bt=dos -bcl=dos)
 elseif(UNIX)
-  target_sources(asciipat PRIVATE spec_nix.cpp)
+  target_sources(asciipat PRIVATE spec_nix.cpp
+  $<IF:$<BOOL:${sound}>,pulse_audio.cpp,sound_stub.cpp>
+  )
   target_link_libraries(asciipat PRIVATE X11::X11 ${XI_LIBRARY} ${PULSEAUDIO_LIBRARY} Threads::Threads)
   target_include_directories(asciipat PRIVATE ${XI_INCLUDE_DIR} ${PULSEAUDIO_INCLUDE_DIR})
 elseif(WIN32)
   target_compile_definitions(asciipat PRIVATE
   "$<$<BOOL:${MSVC}>:NOMINMAX;_CRT_SECURE_NO_WARNINGS>"
   )
-  target_sources(asciipat PRIVATE spec_win.cpp)
-  target_link_libraries(asciipat PRIVATE dsound winmm gdi32 dxguid)
+  target_sources(asciipat PRIVATE spec_win.cpp
+  $<IF:$<BOOL:${sound}>,win_sound.cpp,sound_stub.cpp>
+  )
+  target_link_libraries(asciipat PRIVATE winmm "$<$<BOOL:${sound}>:dsound;gdi32;dxguid>")
 else()
   message(FATAL_ERROR "Unknown platform / cross-compile configuration.
   CMAKE_SYSTEM_NAME: ${CMAKE_SYSTEM_NAME}
