+++ Makefile
@@ -4,7 +4,7 @@ LD       = g++
 LDFLAGS  = -pthread
 LIBS     = -L/usr/X11/lib -lX11 -lXi -lpulse
 OBJS     = manual.o mo3.o unmo3.o stb_vorbis.o conf.o gameover.o inter.o \
-           twister.o game.o temp.o menu.o assets.o spec_nix.o
+           twister.o game.o temp.o menu.o assets.o spec_nix.o pulse_audio.o
 
 .cpp.o:
 	$(CXX) -c $(CXXFLAGS) -o $*.o $<
